# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drawbridge API",
    "version": "0.0.0",
    "description": "The API behind Drawbridge's web UI (docs/PLAN.md §8).\n\nIt's served only to the home network and the VPN, over HTTPS on port 51821. Log in with POST /api/auth/login; the session is a cookie, `__Host-drawbridge`. Every request that changes something needs the `X-Drawbridge` header (any value) and, when it has a body, `Content-Type: application/json`."
  },
...
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/clients": {
      "get": {
        "summary": "List clients",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Every client with its live state, ordered by IPv4 address.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "clients"
        ]
      },
      "post": {
        "summary": "Add a client",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "201": {
            "description": "Created, with the next free addresses and new keys.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResult"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "The new client's URL."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "tags": [
          "clients"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewClient"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, for people."
          }
        }
      },
...
      "Peer": {
        "type": "object",
        "required": [
          "receive_bytes",
          "send_bytes"
        ],
        "description": "A client's live state in the tunnel.",
        "properties": {
          "endpoint": {
            "type": "string",
            "description": "Where the client's packets come from, as address:port."
          },
          "last_handshake": {
            "type": "string",
            "format": "date-time",
            "description": "Absent if the client has never completed a handshake."
          },
          "receive_bytes": {
            "type": "integer",
            "description": "All-time total; resets only when the peer is recreated (a pause and resume, or the tunnel restarting)."
          },
          "send_bytes": {
            "type": "integer",
            "description": "All-time total; resets only when the peer is recreated (a pause and resume, or the tunnel restarting)."
          },
          "session_started_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the client's current connection began. Absent when it has none."
          },
          "session_receive_bytes": {
            "type": "integer",
            "description": "Bytes received during the client's current connection. Absent when it has none."
          },
          "session_send_bytes": {
            "type": "integer",
            "description": "Bytes sent during the client's current connection. Absent when it has none."
          }
        }
      },
      "Client": {
        "type": "object",
        "required": [
          "id",
          "name",
          "enabled",
          "ipv4",
          "ipv6",
          "public_key",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "description": "False while the client is paused."
          },
          "ipv4": {
            "type": "string"
          },
          "ipv6": {
            "type": "string",
            "description": "Empty when IPv6 is off."
          },
          "public_key": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "peer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Peer"
              }
            ],
            "description": "Absent when the client isn't in the tunnel: it's paused, or the tunnel is down."
          }
        }
      },
      "NewClient": {
        "type": "object",
        "required": [
          "name"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "Letters, digits, spaces, and . _ ' -, starting with a letter or digit."
          }
        }
      },
...
      "ClientResult": {
        "type": "object",
        "required": [
          "client"
        ],
        "properties": {
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "warning": {
            "type": "string"
          },
          "apply_failed": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-drawbridge"
      }
    },
    "parameters": {
      "CSRF": {
        "name": "X-Drawbridge",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Any value. Browsers can't send it cross-origin without a CORS preflight, which Drawbridge refuses."
      },
...
    },
...
    "responses": {
      "Invalid": {
        "description": "The request is invalid; the message says why.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotLoggedIn": {
        "description": "Not logged in, or the session ended.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The source isn't on the home network or the VPN, or the request failed the CSRF checks.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "The change conflicts with the current state, such as a name that's taken.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}